Filter comments from statements
This commit updates SQLExecute.run to strip comments from statements.
This is a workaround for a bug where parsed source files that contain
comments can hide special commands. A better solution would refactor
SQLExecute.run and main.execute to use sqlparse.Statement when
determining the type of a statement.

This change was made to present a workaround for a bug found during
normal usage.