Improve path suggestions for source commands
This commit improves the path suggestions provided when specifying a
source file. Suggestions now include the files within the specified
directory. An empty file path (e.g. "source ") now displays the list of
visible files in the current working directory. The list of suggested
paths filters down as a file path is specified.

This change was made to make it easier to locate source files.